requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.0
aioredis==1.3.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==5.0.1
asyncpg==0.30.0
attrs==22.2.0
black==25.1.0
build==1.1.1
CacheControl==0.14.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.8
colorama==0.4.6
coverage==7.0.5
crashtest==0.4.1
crispy-bootstrap4==2024.1
decorator==5.1.1
distlib==0.3.8
Django==4.2
django-cors-headers==4.5.0
django-crispy-forms==2.3
djangorestframework==3.15.2
drf-yasg==1.21.8
dulwich==0.21.7
executing==2.0.1
fastapi==0.115.12
fastapi-cache2==0.2.2
fastjsonschema==2.19.1
filelock==3.13.1
flake8==7.1.1
frozenlist==1.5.0
getenv==0.2.0
greenlet==3.2.0
h11==0.14.0
hiredis==3.1.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
installer==0.7.0
ipython==8.26.0
isort==5.13.2
jaraco.classes==3.3.1
jedi==0.19.1
keyring==24.3.1
matplotlib-inline==0.1.7
mccabe==0.7.0
more-itertools==10.2.0
msgpack==1.0.8
multidict==6.1.0
mypy-extensions==1.0.0
packaging==25.0
parso==0.8.4
pathspec==0.12.1
pendulum==3.0.0
pexpect==4.9.0
pillow==10.4.0
pkginfo==1.10.0
platformdirs==4.3.7
pluggy==1.5.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
prompt_toolkit==3.0.47
propcache==0.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.12.1
pycparser==2.21
pydantic==2.11.3
pydantic_core==2.33.1
pyflakes==3.2.0
Pygments==2.18.0
pyproject_hooks==1.0.0
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==4.0.0
pytest-dotenv==0.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytils==0.4.1
pytz==2024.2
PyYAML==6.0.2
rapidfuzz==3.6.2
redis==5.2.1
requests==2.31.0
requests-toolbelt==1.0.0
setuptools==75.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.46.2
time-machine==2.16.0
tomlkit==0.12.4
traitlets==5.14.3
trove-classifiers==2024.3.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.0
uvicorn==0.34.2
uvloop==0.21.0
virtualenv==20.25.1
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
xattr==1.1.0
yarl==1.17.1
